[Add] Coin UI/ 체력 수정
실제 플레이 중의 체력 값이랑 ui 맞춰줌/ coin ui 구현함

+++ Assets/Scripts/ScriptableObjects/HealthIncreaseTrigger.cs
@@ -5,19 +5,15 @@ using UnityEngine;
 public class HealthIncreaseTrigger : MonoBehaviour
 {
     [SerializeField, Tooltip("How much should the player's health Increase")]
-    private int helathIncreaseAmount = 10;
+    private float helathIncreaseAmount = 10;
 
     [SerializeField]
     private HealthScriptableObject healthManager;
-
-   // public GameEvent healthIncreaseEvent;
 
     public void OnTriggerEnter2D(Collider2D other)
     {
         if(other.CompareTag("Player"))
         {
-           // other.gameObject.GetComponent<HealthScriptableObject>().IncreaseHealth(helathIncreaseAmount);
-           // healthIncreaseEvent.Raise();
             healthManager.IncreaseHealth(helathIncreaseAmount);
             gameObject.SetActive(false);
         }
